Have LightDM use black background
Apparently, the display manager sets the background (wallpaper) for the
X11 environment. If I'm using LightDM, then I'm not using KDE, and if
I'm not using KDE (which is most of the time), then I don't want frills
like a pattern and a NixOS logo where I used to (with SLiM) have just a
blank black screen. Therefore, have LightDM give me a simple black
background.