Vim spellfile: Fix `if` test in setup script
In the snippet of Vim script that instructs Vim to use my spellfile, I
had incorrectly used `!&spellfile` in an `if` statement's condition
expression to check whether the 'spellfile' Vim option's value was the
empty string, when the correct form would be `&spellfile == ''`.

I had used this incorrect form after insufficiently testing the
application of Vim script's `!` (logical negation) operator to
strings, such that I knew that `!''` (the logical negation of the
empty string) evaluates to one (the "true" value), but apparently
without checking what the logical negation of a non-empty string
evaluates to.

Apparently, Vim's logical negation operator, when applied to a string,
tries to parse a number from the string with what would seem to be
`atoi` or a similar function (that parses as much number as it can
without minding the whole string not being a number), then applies
itself to the parsed number.

Thus, `!&spellfile` is incorrect for testing whether `&spellfile` is
the empty string.

This change fixes this issue.