Revert "Drop dmenu, i3status from system packages"
This reverts commit 5b1df50c40762652ced2fc5210ab5e4c0d65a9f5.

While i3 does declare `dmenu` and `i3status` as dependencies, the
`dmenu` package is broken such that, when it is omitted from the set
of system packages, i3's invocation of dmenu results in a menu devoid
of menu entries.

I have written a patch to fix this issue and filed a GitHub pull
request [1] to nixpkgs upstream for said patch.

Pending resolution of said pull request, I'm reverting my commit that
dropped `dmenu` from the set of system packages.

[1]: <https://github.com/NixOS/nixpkgs/pull/17709>