Ensure X11 virtual terminal gets set to RAW mode
It seems that as of a recent NixOS update, the virtual terminal used
by X11 (typically `/dev/tty7`) is not set to RAW keyboard mode as it
should be, resulting in presses of Alt-modified arrow keys and
function keys being interpreted by the kernel (as
virtual-terminal-switching commands) rather than being left solely to
X (to which they are additionally passed).

This commit adds a systemd unit that detects which virtual terminals
are used by running X servers, and sets those virtual terminals to RAW
keyboard mode.