sw-modules: Allow installing not to systemPackages
In the software module system, support installing packages such that
they are kept in the Nix store but not added to the set of system
packages.

This allows for, e.g., programming language tool-chains, which are
inappropriate to keep in the set of system packages, to be kept around
for access via `nix-shell` without risk of their being collected as
garbage by `nix-collect-garbage`.