VMs: disable tmpOnTmpfs, saving 9P dirs
NixOS virtual machines (the sort started with commands with names of
the form `run-<hostname>-vm`) provide two 9P mount-points for
exchanging data with the host system: `/tmp/shared` and `/tmp/xchg`.

If the NixOS option `boot.tmpOnTmpfs` is set to `true` in a virtual
machine's configuration, however, that option will cause a tmpfs to be
mounted at `/tmp` *after* the two 9P directories are created in
`/tmp`, such that the tmpfs is mounted over the 9P directories,
rendering them inaccessible.

I have filed [an issue report][1] against nixpkgs re this issue,
pending the resolution of which report, this patch disables
`boot.tmpOnTmpfs` for such virtual machines.

[1]: <https://github.com/NixOS/nixpkgs/issues/21490>