Linux kernel: Upgrade from LTS to latest stable
Upgrade my Linux kernel package from the long-term support (LTS)
version 4.9 to the latest stable version 4.12.

The reason for this change is that Linux 4.9 no longer builds
successfully with the NixOS hardened kernel configuration.

I had been using the LTS version because ZFS worked with it but not
(according to NixOS) with the latest stable, but now NixOS is okay
with me using ZFS with latest stable Linux.