Chromium: Disable default browser check
Wrap Chromium in a shell script passing it the command-line argument
`--no-default-browser-check`, telling it not to complain that it's not
the default Web browser.

I would let it be the default browser, but the only notion of "default
browser" my desktop-environment-less system has, as far as I am aware,
is the environment variable `BROWSER`, about which Chromium doesn't
seem to care.