Give up on NetworkManager
I had tried switching from wpa_supplicant to NetworkManager because I
failed to get wpa_supplicant to connect to a Wi-Fi network that uses
WPA-Enterprise and to which it sometimes would be convenient for me to
connect. However, NetworkManager, it turns out, not only doesn't
connect to that network either, it also doesn't connect to my home
Wi-Fi network, making it useless to me.