lib/date-and-time: Fix zero-padding of hour
Previously, the functions converting an hour number in local time to a
string representation of said hour number in target time would
zero-pad the output to two digits iff the *input* hour number was less
than 10, which was wrong. This patch fixes these functions to zero-pad
iff the *output* hour number is less than 10.